requirement.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
streamlit==1.37.1
pyyaml==6.0.2
ccxt==4.4.65
pydantic==2.8.2
requests==2.32.3
tenacity==8.5.0